Fixed issue where newly selected settings were not highlighted
While messing around with the original colour scheme, before the initial commit to GitHub, I set the background colour for newly selected settings to one that did not distinguish it at all from the div it was part of. The background colour for newly selected settings has now been changed to fix this.