Fix "generation finished" message not being shown
This was caused by the recent changes to the preloading code. It caused
the first Pokémon to be skipped, and at the end of the generation the
returned number was `undefined` instead of `-1`. This meant the end of
generation message was never shown, and new numbers were generated
immediately instead. This commit fixes the issue by generating the
numbers correctly.

I've also renamed some of the variables in this commit to hopefully
make it a bit more clear what they actually are.

Fixes #43