Rewrite entire codebase to use Preact and Redux
This is pretty much a complete rewrite of the game. Moving to Preact and
Redux will make it significantly easier to add new features in the
future, such as advanced stats or other game modes.

It should also fix a whole class of bugs that were always popping up
relating to the UI being out of sync with the state. All the state is
now stored in Redux, meaning the UI is always automatically in sync with
the state. The entire state is now also persisted, meaning users can
resume games across reloads, maintaining streaks and continuing with the
same numbers they had before.

There are a few things still to do before releasing this:
 - When generations are changed, add and remove numbers to the existing
   array if possible, without affecting the current Pokémon
 - Add mobile stats summary
 - Lots more testing!