Only apply generation and difficulty changes after an answer
This matches the existing behaviour. I originally planned to do
something more clever where the numbers array is modified in place
without shuffling, but it ended up remarkably complex with a number of
edge cases. This approach is more simple, and will not be a surprise
to existing users.

This commit also adds a new counter which shows the user's progress
through their current set of numbers.